Simplify sudoers rules
With the 2562b1a change to one shared login user it doesn't make as
much sense to have individual per zone sudoers rules. Also, the rndc
and knotc subcommands are narrow enough that I really don't see any
real room for wildcard abuse.

On the upside, this simplification takes away the need to rerun the
_szh-sudoers_ command any time a zones gets added to the
_zone-handler.yaml_ config file.